Finished converting remaining null cases for gases to being nonnull, there is a slight chance I missed some. Still need to finish reviewing converting infuse types to nonnull and making infusion tank store an infusion stack

+++ src/main/java/mekanism/client/sound/PlayerSound.java
@@ -104,7 +104,7 @@ public abstract class PlayerSound extends LocatableSound implements ITickableSou
 
     @Override
     public float getVolume() {
-        return (float) (super.getVolume() * MekanismConfig.client.baseSoundVolume.get());
+        return super.getVolume() * MekanismConfig.client.baseSoundVolume.get();
     }
 
     public enum SoundType {
